Add preamble.sql -- util for help loading schemas
It's useful to be able to embed commented-out COMMENT statements inside
TABLEs, for example, and then use a psql shell escape to apply those
COMMENTs.  This requires having the PG conninfo/URI, for example, and
also the name of the file being loaded.

We should really contribute a patch to add builtin psql variables like:

 - PG_CONNINFO

   The conninfo given to psql.

 - FNAME

   The file name given to psql with -f <file>.

 - FDIRNAME

   The directory containing the file given to psql with -f <file>.