COMMIT TRIGGERs: add BEGIN/CONNECT, make secure
This adds a prototype of BEGIN and CONNECT triggers as well, and uses
this internally to detect the use of SET CONSTRAINTS .. IMMEDIATE and
raises an exception if that would break commit triggers.

Also, if a connect/begin/commit trigger raises an exception which is
then caught, and the transaction ultimately continues, then those
triggers will be re-run at the next opportunity.